Fixed bug #0000335: If $this->dont_use_exec_commands was set, it was impossible to delete any directories, which is now possible again. It's still impossible to delete recursive direcories though.
git-svn-id: https://svn.typo3.org/TYPO3v4/Core/trunk@578 709f56b5-9817-0410-a4d7-c38de5d9e867

+++ t3lib/class.t3lib_extfilefunc.php
@@ -413,12 +413,12 @@ class t3lib_extFileFunctions extends t3lib_basicFileFunctions	{
 				} else $this->writelog(4,1,112,'You are not allowed to delete files','');
 				// FINISHED deleting file
 
-			} elseif (@is_dir($theFile) && !$this->dont_use_exec_commands) {	// if we're deleting a folder
+			} elseif (@is_dir($theFile)) {	// if we're deleting a folder
 				if ($this->actionPerms['deleteFolder'])	{
 					$theFile = $this->is_directory($theFile);
 					if ($theFile)	{
 						if ($this->checkPathAgainstMounts($theFile))	{	// I choose not to append '/' to $theFile here as this will prevent us from deleting mounts!! (which makes sense to me...)
-							if ($this->actionPerms['deleteFolderRecursively'])	{
+							if ($this->actionPerms['deleteFolderRecursively'] && !$this->dont_use_exec_commands)	{
 									// No way to do this under windows
 								$cmd = 'rm -Rf "'.$theFile.'"';
 								exec($cmd);		// This is a quite critical command...
